Parse % as value if prior token is not variable
A valid fortran namelist can set a variable equal to a special token as
a string without delimiters, e.g.

&nml
   x = %
/

sets x='%'.  This is difficult to parse in f90nml since we rely on
tokens to identify and distinguish variables.

This patch does not solve the issue, but partially fixes the problem by
identifying that '=' is not a variable name, and therefore '%' should be
interpreted as a value.

Subsequent checks do not correctly read '%' as a value, but it does
prevent corruption of subsequent values.